requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
